+++ Release-Notes.md
@@ -1,6 +1,11 @@
 
 # PsdKit Release Notes
 
+## v0.0.3
+
+- Fix #1 (for now, write enums as strings)
+- Treat `SwitchParameter` as Boolean.
+
 ## v0.0.2
 
 - Do not sort dictionary keys in `ConvertTo-Psd`.